Components were extracted into a separate directory.
