+++ examples/pong/components/mod.rs
@@ -1,0 +1,7 @@
+mod ball;
+mod paddle;
+
+pub use self::{
+    ball::Ball,
+    paddle::{Paddle, Side},
+};
