+++ examples/pong/components/paddle.rs
@@ -1,0 +1,31 @@
+use amethyst::ecs::{Component, DenseVecStorage};
+
+use crate::{PADDLE_HEIGHT, PADDLE_VELOCITY, PADDLE_WIDTH};
+
+#[derive(PartialEq, Eq)]
+pub enum Side {
+    Left,
+    Right,
+}
+
+pub struct Paddle {
+    pub velocity: f32,
+    pub side: Side,
+    pub width: f32,
+    pub height: f32,
+}
+
+impl Paddle {
+    pub fn new(side: Side) -> Paddle {
+        Paddle {
+            velocity: PADDLE_VELOCITY,
+            side,
+            width: PADDLE_WIDTH,
+            height: PADDLE_HEIGHT,
+        }
+    }
+}
+
+impl Component for Paddle {
+    type Storage = DenseVecStorage<Self>;
+}
